Move target comparison to internal header file
