+++ cpp/src/Ice/LocatorInfo.cpp
@@ -2,7 +2,6 @@
 
 #include "LocatorInfo.h"
 #include "EndpointI.h"
-#include "Ice/Comparable.h"
 #include "Ice/LocalExceptions.h"
 #include "Ice/Locator.h"
 #include "Ice/LoggerUtil.h"
