+++ cpp/src/Ice/msbuild/ice/ice.vcxproj.filters
@@ -808,9 +808,6 @@
     <ClInclude Include="..\..\..\..\include\Ice\CommunicatorF.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\..\..\include\Ice\Comparable.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\..\..\include\Ice\Config.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -971,6 +968,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\..\..\..\include\Ice\StringConverter.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\..\include\Ice\TupleCompare.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\..\..\..\include\Ice\UserException.h">
