+++ cpp/test/Ice/proxy/AllTests.cpp
@@ -1,5 +1,6 @@
 // Copyright (c) ZeroC, Inc.
 
+#include "../../src/Ice/TargetCompare.h"
 #include "Ice/Ice.h"
 #include "Ice/Locator.h"
 #include "Ice/Router.h"
