+++ slice/Ice/Version.ice
@@ -9,7 +9,7 @@
 [["cpp:no-default-include"]]
 [["cpp:no-stream"]]
 [["cpp:include:Ice/Config.h"]]
-[["cpp:include:Ice/Comparable.h"]]
+[["cpp:include:Ice/TupleCompare.h"]]
 [["cpp:include:cstdint"]]
 [["cpp:include:ostream"]]
